requirements.txt
streamlit>=1.38.0
networkx>=3.2
pyvis>=0.3.2
plotly>=5.22.0
requests>=2.31.0
python-dotenv>=1.0.1
rapidfuzz>=3.8.1
Pillow>=10.0.0
ImageHash>=4.3.1
instaloader>=4.10.1
praw>=7.7.1
google-api-python-client>=2.149.0
